mtrl: pin Material/CustomMaterial OnFile fields to explicit-width types
Replace bare long/short with int32_t/int16_t in CustomMaterialOnFile and
MaterialOnFile. No layout change on x86; static_asserts at sizeof(...)==8
and ==20 verify.

Project 1 audit step 3.

+++ inc/mtrl.h
@@ -15,14 +15,16 @@
 #ifndef MTRL_H
 #define MTRL_H
 
+#include <cstdint>
+
 using namespace BRender;
 
 // CustomMaterial_CMTL on File
 struct CustomMaterialOnFile
 {
-    short bo;
-    short osk;
-    long ibset; // which body part set this CustomMaterial_CMTL attaches to
+    int16_t bo;
+    int16_t osk;
+    int32_t ibset; // which body part set this CustomMaterial_CMTL attaches to
 };
 static_assert(sizeof(CustomMaterialOnFile) == 8, "CustomMaterialOnFile on-disk layout drift");
 const ByteOrderMask kbomCmtlf = 0x5c000000;
@@ -30,8 +32,8 @@ const ByteOrderMask kbomCmtlf = 0x5c000000;
 // material on file (Material_MTRL chunk)
 struct MaterialOnFile
 {
-    short bo;            // byte order
-    short osk;           // OS kind
+    int16_t bo;          // byte order
+    int16_t osk;         // OS kind
     br_colour brc;       // RGB color
     br_ufraction brufKa; // ambient component
     br_ufraction brufKd; // diffuse component
